refactor: create RefitPipeline orchestrator and refactor RefitGenerator to thin facade
